requirements.txt
django-admin-interface
django-colorfield
django-versatileimagefield
django-money
django-tinymce
django-import-export
django-crispy-forms
crispy-bootstrap5
django-tables2
django-filter
django-registration-redux
django-extensions
django_compressor
drf_yasg
djangorestframework-simplejwt
django-rosetta
python-decouple
nltk
onesignal-python-api
